[ROCm/gfx1250] Fix fp8 KV cache MLA attention: upcast fp8 K to bf16 in QK dots
On gfx1250, triton `tl.dot(fp8, fp8)` returns garbage (~1e34+) once the
contraction dim K >= 128 (K=64 is correct; bf16 tl.dot is correct at all K) —
a gfx1250-specific fp8-MFMA codegen issue. The DeepSeek MLA nope QK dot has
K=512, and the triton MLA decode/extend kernels downcast the query to fp8 to
match an fp8 KV cache (`q.to(K_Buffer.dtype)`), so with `--kv-cache-dtype
fp8_e4m3` the QK dot produces garbage -> softmax(inf) -> NaN -> degenerate
decode (empty/whitespace output, GSM8K ~0).

Fix: keep q in its native (bf16) dtype and upcast the fp8 K to q's dtype in the
dot instead of downcasting q to fp8. Applies to:
  - decode_attention.py::_fwd_grouped_kernel_stage1 (decode reading fp8 KV)
  - extend_attention.py::_fwd_kernel and _fwd_kernel_unified prefix loops
    (fire when radix cache reuses a shared prefix -> prefill reads the fp8 KV)
Both nope and rope (pe) dots are upcast. No-op for a bf16 KV cache.

Validated on gfx1250 (DeepSeek-R1-0528-MXFP4, full recipe: radix + cuda-graph):
fp8 KV GSM8K 1319Q 0.949 == bf16 KV 0.951, full speed (~182 tok/s). Halves KV
cache memory. gfx950 (same triton code) is unaffected (its fp8 dot at K=512 is
correct).

+++ python/sglang/kernels/ops/attention/extend_attention.py
@@ -421,7 +421,12 @@ def _fwd_kernel(
                 mask=(mask_n[None, :]) & (mask_d[:, None]),
                 other=0.0,
             )
-            qk = tl.dot(q.to(k.dtype), k)
+            # gfx1250: triton tl.dot(fp8, fp8) returns garbage (~1e34+) for contraction
+            # dim K>=128 (K=64 ok). This prefix read fires when a radix-cache prefix is
+            # reused (prefill reads the cached fp8 KV), and the MLA nope dot has K=512,
+            # so we must upcast the fp8 K to q's dtype and dot in bf16 rather than
+            # downcasting q to fp8. No-op for a bf16 cache. (Do NOT revert to q.to(fp8).)
+            qk = tl.dot(q, k.to(q.dtype))
             if BLOCK_DPE > 0:
                 if PAGE_SIZE == 1:
                     offs_kpe = (
@@ -441,7 +446,7 @@ def _fwd_kernel(
                     mask=mask_n[None, :],
                     other=0.0,
                 )
-                qk += tl.dot(qpe.to(kpe.dtype), kpe)
+                qk += tl.dot(qpe, kpe.to(qpe.dtype))
             qk *= sm_scale * k_scale
 
             if logit_cap > 0:
@@ -996,7 +1001,12 @@ def _fwd_kernel_unified(
                 other=0.0,
             )
 
-            qk = tl.dot(q.to(k.dtype), k)
+            # gfx1250: triton tl.dot(fp8, fp8) returns garbage (~1e34+) for contraction
+            # dim K>=128 (K=64 ok). This prefix read fires when a radix-cache prefix is
+            # reused (prefill reads the cached fp8 KV), and the MLA nope dot has K=512,
+            # so we must upcast the fp8 K to q's dtype and dot in bf16 rather than
+            # downcasting q to fp8. No-op for a bf16 cache. (Do NOT revert to q.to(fp8).)
+            qk = tl.dot(q, k.to(q.dtype))
             if BLOCK_DPE > 0:
                 if PAGE_SIZE == 1:
                     offs_kpe = (
@@ -1016,7 +1026,7 @@ def _fwd_kernel_unified(
                     mask=mask_n[None, :],
                     other=0.0,
                 )
-                qk += tl.dot(qpe.to(kpe.dtype), kpe)
+                qk += tl.dot(qpe, kpe.to(qpe.dtype))
 
             qk *= sm_scale_withk
